third_party/WebKit/LayoutTests/platform/win/http/tests/w3c/webperf/submission/Intel/user-timing/test_user_timing_method_exist-expected.txt

This is a testharness.js-based test.
PASS window.performance is defined 
PASS performance.mark is supported! 
PASS performance.clearMarks is supported! 
PASS performance.measure is supported! 
PASS performance.clearMeasures is supported! 
PASS performance.getEntries is supported! 
PASS performance.getEntriesByType is supported! 
PASS performance.getEntriesByName is supported! 
Harness: the test ran to completion.
